requirements.txt
opencv-python~=4.11.0.86
numpy<3
pynetworktables~=2021.0.0
ultralytics~=8.3.158
Flask~=3.1.1
numba~=0.60.0
tensorflow~=2.19.0

customtkinter~=5.2.2
torch~=2.7.1
tqdm~=4.67.1
pip~=25.1
pillow~=11.2.1
filelock~=3.18.0
clearml~=2.0.0

pygrabber~=0.2
comtypes~=1.4.11

imutils~=0.5.4
flask_socketio~=5.5.1

pupil-apriltags~=1.0.4